test-requirements.txt
pytest>=7.2.2
python-dotenv>=1.1.0
playwright>=1.52.0
pytest-rerunfailures>=12.0